Organize, reset, modularize
Beginning some good work on thresholding techniques. Demos using several
